requirements.txt
certifi==2019.11.28
Django==2.2.7
numpy==1.17.4
pandas==0.25.3
pandas-datareader==0.8.1
psycopg2==2.8.4
pyparsing==2.4.5
python-dateutil==2.8.1
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.7
django_cron==0.5.1
djangorestframework==3.11.0
django-rest-swagger==2.2.0
celery==4.2.0
django-rest-auth==0.9.5
django-allauth==0.41.0